fixe

+++ controler/teacher.manage.ctrl.php
@@ -13,6 +13,7 @@ if(!isset($_SESSION["id"])) {
 } else if($_SESSION["roleid"] != 2){
     include(__DIR__."/../controler/landing.ctrl.php");
 }
+
 elseif(!isset($_POST["stats"])){
 
 
@@ -23,58 +24,63 @@ elseif(!isset($_POST["stats"])){
     }
 
     $currentClass = ClassGroup::getClassGroupFromId($currentClassId);
-
     $teacher = Teacher::readTeacher($_SESSION['id']);
+    if(isset($_POST["modif"])){
+        $view->assign("class",$currentClass);
+        $view->assign("teacher", $teacher);
+        $view->display("teacher.modif.ctrl.php");
+    }else{
+        if(isset($_POST['create'])){
+            $newClass = new ClassGroup($teacher);
+            $newClass->create();
+        }
 
-    if(isset($_POST['create'])){
-        $newClass = new ClassGroup($teacher);
-        $newClass->create();
-    }
-
-    $classList = ClassGroup::getClassGroupsFromTeacher($teacher);
-
-    if(isset($_POST['deleteGroup'])){
-        $currentClass->delete();
         $classList = ClassGroup::getClassGroupsFromTeacher($teacher);
-        $currentClass = null;
-    }
 
-    if(count($classList) == 0){
-        $students = [];
-        $className = "Pas de classe pour le moment ! ";
-        $classList = [];
-        $code = "";
-    }
-    else{
-        if($currentClass == null){
-            $currentClass = $classList[0];
-        }
-
-        if(isset($_POST['delete'])){
-            $studentToDel = Student::readStudent($_POST['delete']);
-            $currentClass->removeStudent($studentToDel);
-        }
-
-        if(isset($_POST['updateName'])){
-            $currentClass->setName($_POST['className']);
+        if(isset($_POST['deleteGroup'])){
+            $currentClass->delete();
             $classList = ClassGroup::getClassGroupsFromTeacher($teacher);
+            $currentClass = null;
         }
 
-        $students = $currentClass->getStudents();
-        $className = $currentClass->getName();
-        $code = "Code : " . $currentClass->getCode();
 
-        
+        if(count($classList) == 0){
+            $students = [];
+            $className = "Pas de classe pour le moment ! ";
+            $classList = [];
+            $code = "";
+        }
+        else{
+            if($currentClass == null){
+                $currentClass = $classList[0];
+            }
+
+            if(isset($_POST['delete'])){
+                $studentToDel = Student::readStudent($_POST['delete']);
+                $currentClass->removeStudent($studentToDel);
+            }
+
+            if(isset($_POST['updateName'])){
+                $currentClass->setName($_POST['className']);
+                $classList = ClassGroup::getClassGroupsFromTeacher($teacher);
+            }
+
+            $students = $currentClass->getStudents();
+            $className = $currentClass->getName();
+            $code = "Le code de la classe : " . $currentClass->getCode();
+
+
+        }
+
+
+
+        $view->assign("currentClass",$currentClass);
+        $view->assign("code",$code);
+        $view->assign("students",$students);
+        $view->assign("className",$className);
+        $view->assign("classList",$classList);
+        $view->display("teacher.manage.view.php");
     }
-
-    
-    
-    $view->assign("currentClass",$currentClass);
-    $view->assign("code",$code);
-    $view->assign("students",$students);
-    $view->assign("className",$className);
-    $view->assign("classList",$classList);
-    $view->display("teacher.manage.view.php");
 }
 else{ //Gestion stats élève
     $student = Student::readStudent($_POST["stats"]);
